Close manifest-update outcome enum, split output tests, add assert messages (#97)
Address a batch of review findings on the bump command and its tests.

- bump.py: replace the two-boolean `_CrateManifestUpdate` dataclass with a
  closed `_CrateManifestOutcome` enum (SKIPPED / UPDATED / UNCHANGED) so the
  both-true state is unrepresentable; `_process_crate_manifests` selects on it
  with match/case. Remove the redundant `_update_crate_manifest` wrapper (a
  pass-through once the outcome is an enum) and point its sole caller at
  `_apply_crate_manifest_update`.
- Split test_bump_result_formatting.py: the change-description, grammar, and
  snapshot result-message tests move to a new test_bump_output_formatting.py
  (with the snapshot file renamed accordingly), leaving the original focused on
  `_apply_crate_manifest_update` and its fixtures. Both modules are now well
  under the 400-line cap.
- Cap the I/O-heavy manifest-selection Hypothesis test at max_examples=20,
  matching the repo's established pattern.
- Add descriptive failure messages to the bare asserts across
  test_bump_selectors, test_bump_toml_helpers, test_bump_manifest_writing,
  test_bump_context_properties, and test_bump_output_formatting; the manifest
  property helper now names the crate and reports expected vs actual.

No behavioural change to the bump command; the enum carries the same
skipped/updated/unchanged information the boolean pair did.
